requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
anthropic>=0.18.0
gunicorn>=23.0.0
PyJWT[crypto]>=2.8.0
